reorganize and reformat
