+++ deepmd/cluster/__init__.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 # @File    : __init__.py.py
-# @Time    : 1/23/2021 4:52 PM
+# @Time    : 1/23/2021 9:20 PM
 # @Author  : Zavier Cai
 # @Email   : caizefeng18@gmail.com
-# TODO: modify importing paths of `tests` directory
