+++ deepmd/cmake/cmake_lammps.cmake.in
@@ -1,8 +1,10 @@
 set (LMP_INSTALL_PREFIX "@CMAKE_BINARY_DIR@/USER-DEEPMD")
 file(READ "@CMAKE_BINARY_DIR@/lmp/lammps_install_list.txt" files)
 string(REGEX REPLACE "\n" "" files "${files}")
-
+message(${files})
 foreach (cur_file ${files})
+
+
   file (
     INSTALL DESTINATION "${LMP_INSTALL_PREFIX}" 
     TYPE FILE
@@ -18,6 +20,6 @@ file (
 
 file (
   INSTALL DESTINATION "${LMP_INSTALL_PREFIX}" 
-  TYPE FILE 
+  TYPE FILE
   FILES "@CMAKE_BINARY_DIR@/lmp/pair_nnp.h"
 )
