+++ deepmd/scripts/CMakeLists.txt
@@ -1,5 +1,5 @@
 install(
   FILES		freeze.py config.py 
-  DESTINATION	deepmd/
+  DESTINATION	deepmd/scripts
 )
