+++ deepmd/train/DescrptSeAR.py
@@ -1,10 +1,10 @@
 import numpy as np
-from deepmd.env import tf
-from deepmd.common import ClassArg
+from deepmd.train.env import tf
+from deepmd.train.common import ClassArg
 
-from deepmd.DescrptSeA import DescrptSeA
-from deepmd.DescrptSeR import DescrptSeR
-from deepmd.env import op_module
+from deepmd.train.DescrptSeA import DescrptSeA
+from deepmd.train.DescrptSeR import DescrptSeR
+from deepmd.train.env import op_module
 
 class DescrptSeAR ():
     def __init__ (self, jdata):
