+++ deepmd/train/Loss.py
@@ -1,9 +1,9 @@
 import numpy as np
-from deepmd.env import tf
-from deepmd.common import ClassArg, add_data_requirement
+from deepmd.train.env import tf
+from deepmd.train.common import ClassArg, add_data_requirement
 
-from deepmd.RunOptions import global_cvt_2_tf_float
-from deepmd.RunOptions import global_cvt_2_ener_float
+from deepmd.train.RunOptions import global_cvt_2_tf_float
+from deepmd.train.RunOptions import global_cvt_2_ener_float
 
 class EnerStdLoss () :
     def __init__ (self, jdata, **kwarg) :
