+++ deepmd/train/doc.py
@@ -1,4 +1,4 @@
-from deepmd.argcheck import gen_doc
+from deepmd.train.argcheck import gen_doc
 
 def doc_train_input(args):
     doc_str = gen_doc(make_anchor=True)
